pyproject.toml
[project]
name = "mms-general"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.13.0",
    "email-validator>=2.3.0",
    "emails>=0.6",
    "fastapi[standard]>=0.135.2",
    "gitpython>=3.1.46",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "rich>=14.3.3",
    "sentry-sdk[fastapi]>=2.56.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "streamlit>=1.55.0",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
    "uvicorn[standard]>=0.42.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
    "types-passlib>=1.7.7.20260211",
    "types-python-jose>=3.5.0.20250531",
]
